refactor: share progress view state helpers
